requirements.txt
loguru==0.7.2
googletrans==4.0.2
pandas
bottleneck
openpyxl==3.1.2
requests==2.31.0
h2==3.2.0
httpx==0.28.1
python-docx
pytest